Tests: Undo integration test is now performed synchronously.
This results with quicker testing. Also gets around the issue https://github.com/ckeditor/ckeditor5/issues/6214. I didn't dive to find out why exactly this was happening, generally the cause of it was that during `await wait()` JS interpreter would trigger a selection change event, which would be picked by our observers and set the selection back to dummy paragraph that is before the image.